remove old sketch; rename sketch2
