--- sketch2/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "sketch2"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-doctest = false
-
-[dev-dependencies]
-tokio-test = "0.4.3"
-
-[workspace]
